montana: Post-Boot Set interactive on both clusters
This commit will make interactive be set on both clusters post boot. This was because we used potters post boot ramdisk post boot script and potter is a single cluster . This commit also sets the correct  freqs and the hispeed freq for the big cluster , This should fix other freqs not getting used and improve battery because the freq isnt always at max.

Change-Id: I740c05d1626af378adde910c60f9f14411804342